JVMCBC-557: Avoid accessing event loop for failed channel
Motivation
----------
This is a regression from JVMCBC-500, which only affected
the 2.6 prerelease branch and no stable releases.

With the netty upgrade we had to introduce a FailedChannel
sine null was no longer accepted. The FailedChannel though
does not have an event loop attached to it. As a result the
if check in the reconnect logic returned true since there
was a channel, but it had no event loop attached and as a result
the reconnect attempt failed.

Modifications
-------------
To remedy this situation, this changeset avoids the overly
complicated if/else logic on using the event loop of the current
channel and right away always picking a new one from the IO pool.

This makes this codepath in general cleaner and less error prone.

Result
------
The retry attempts are not disrupted by a null event loop if the
safeguard triggers on a connect attempt.

Change-Id: I4224f243fe8e876e3c332474ba24c69e654cf8cb
Reviewed-on: http://review.couchbase.org/96435
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>