JVMCBC-556: Distinguish between nodes on the same physical host.
Motivation
----------
This changeset makes it possible that multiple nodes are exposed
via the same physical host, but just on different port ranges.

This is only possible right now via the alternate names config.

Modifications
-------------
To make this work, we need to distinguish between the logical
node and the physical host name which is used to actually connect
the socket.

So throughout the code we still everywhere work with the logical
hostname (the "internal" one) but iff the alternate config is used
we pass this alternate into the node layer so it can propagate
down through the service and the endpoint.

Since at the node layer this "shift" happens from logical
to physical the node code is extended to log both so it's
easier to debug down the road.

Also note that this changeset reverts changes to the NodeInfo
which have been made previously but are actually not working
as expected. We also feed the proper service ports down so
that the correct ports are used only when alternate names
are actually used.

Result
------
If an alternate hostname should be used it is now propagated
down into the service and endpoint layer but the logical
hostname is still used to manage the nodes themselves and
to lookup vbuckets and so forth.

Change-Id: I040862e26e5c2ad41deac0e3a26d7b7e885a9b2b
Reviewed-on: http://review.couchbase.org/97006
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>