JVMCBC-564: Make sure kv service is only enabled if in nodes list.
Motivation
----------
The SDK needs to iterate the nodesExt as part of the config because
some global services (i.e. N1QL) are not to be found in the
traditional "nodes" list of the config. With KV, this turns out
to be a problem during rebalance-in scenarios.

Especially during delta-node recovery and memcached buckets, the
problem amplifies because nodesExt holds all nodes that are part
of the cluster with their services, while the nodes list only
holds the nodes that are currently active for the given bucket.

If the SDK only uses nodesExt (which it does now), it tries to
connect to kv buckets immediately as they become available, even
if they are not active for this bucket yet. For couchbase and
ephemeral buckets this is not a big issue because there the
movement of vbuckets gradually moves over the load, but with
memcache buckets as soon as the node pops up in the config
it is part of the ketama ring and ops are dispatched to it
right away.

Modifications
-------------
To remedy this situation, and also to keep the change scope
as minimal as possible, the part of the config loading where
we iterate nodesExt now explicitly checks if a corresponding
entry in the nodes list is available - and if not, it removes
the KV service from the service list. This will make sure that
only nodes that are also in the nodes list will end up in the
ketama ring and therefore not have ops dispatched too early.

The astute reader might immediately ask him/herself if the
view service should also be disabled because it is colocated
with kv. And the answer is - it likely should, but to keep the
scope small this change is not made. The good news is that
this doesn't really matter because the code somewhere else
only dispatches the op to a view node if there is at least
one vbucket active on that node, so views won't be affected
by this issue (and memcached buckets do not have views
anyways).

A test config has been added from the reproduction system
to make sure this won't surface again.

Result
------
This changeset avoids connecting to kv nodes too early
(both memcache and couchbase, but much more important
for the former) which in turn prevents INVALID_ARGUMENTS
responses, failing ops and disconnects.

Change-Id: I844c0ec2d95701a869b14f811034b6e7bd0f4eae
Reviewed-on: http://review.couchbase.org/99089
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>