JVMCBC-566: Only continuously refresh for CarrierRefresher
Motivation
----------
Previously every refresh interval a signal would be sent from
the CarrierRefresher into the ConfigurationProvider to refresh
all refreshers.

Since only the CarrierRefresher actually did perform a refresh,
this turned out to be a noop on the HttpRerfresher. Now that
we are planning to add refreshing capabilities to the
HttpRefesher as well (but not all the time like with carrier)
the scope needs to change.

Modifications
-------------
This changeset is technically a "noop", because instead of doing
the indirection we now directly signal the refresh interval to
the CarrierRefresher. Nothing is changed here, but once the
HttpRefresher gets refresh capabilities, it is not pinged all
the time.

The reason why we do not want the HttpRefresher to be pinged
all the time is that it already has a streaming config attached
and only needs to run the refresh command if there is explicit
reason from other heuristics (i.e. a socket being closed).

Result
------
Continuous refreshing now only affects the CarrierRefresher
(as before) but the HttpRefresher is not affected in the future.

Change-Id: I4a373f9ee0122835e81ca42f83ceb77e2d6d8799
Reviewed-on: http://review.couchbase.org/99599
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>