JVMCBC-565: Proactively fetch config from http if signaled
Motivation
----------
Previously, when an http stream is attached to fetch a configuration
it has been trusted to always send a proper one in a short timeframe.
When the socket has been closed a new stream would be opened.

This does not account for the fact that a socket can end up in an
half/closed (dead) state and the client never knows that it has been
closed upon it. Since a dead socket is not distinguishable from a
socket where no config stream events are sent over, a new mechanism
needs to be added in order to return to a healthy state.

Modifications
-------------
Other sections of core-io already ask the config manager to grab
a new configuration because of specific heuristics (i.e. a socket
has been closed).

This code adds the behavior to the HttpRefresher which is very
similar to the CarrierRefresher. In fact, most of the code has
been copied 1:1. In a future change this will be refactored into
the common parent but for the sake of clarity we are not touching
the Carrier refresher for now (would be too confusing to review).

As soon as we get our hands on a new configuration the reconfig
logic of the SDK will jump in and clean up / restore a proper
state.

Note that this relies on an external signal from the heuristics,
which means it will go undetected until there is traffic. One
could argue that if there is no traffic also no impact will be
seen. The other approach that has been considered reopened
an idle http connection every x second interval, but this
heavily penalizes the common case for this edge case
scenario.

Result
------
We now properly handle a half/closed (dead) tcp socket where
the streaming connection is attached to.

Change-Id: I1f3a9716fa013d249fe524e3fbc064f136ca7abf
Reviewed-on: http://review.couchbase.org/99600
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>