JVMCBC-567: Properly close ConfigEndpoints when service disconnects
Motivation
----------
While working on a related code-change, it was found that when an
endpoint disconnects it is properly removed from the states, but
when this service in total is disconnected (i.e. if the node is
removed) then the service sticks around, potentially leaving endpoints
connected when they shouldn't be.

Modifications
-------------
This changeset borrows some logic from the pooled service to make sure
that on-demand endpoints are properly managed and if a disconnect
command comes along they are also correctly shudown. This triggers
side effects in certain outstanding requests, for example it closes
http streaming connections properly.

Result
------
The ConfigService (the only one using the on demand service as of
now) correctly handles node disconnects during reconfiguration.

Change-Id: Ic2e4490e3f24c7a38b09d9cffa13a2be1f7cbb74
Reviewed-on: http://review.couchbase.org/99601
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>