JVMCBC-576: Allow to configure SSL protocol via property.
Motivation
----------
This changeset makes it possible to modify the TLS protocol/version
used, which might be needed for example if Java 7 is used and TLS 1.2
needs to be forced onto the client.

Modifications
-------------
Since this is not a broadly requested feature, it is for now only
configurable through a system property. Should there be more demand
we can make this configurable on the environment in later releases
as well.

System property is "com.couchbase.sslProtocol" which defaults to
"TLS" and values must follow the spec as outlined in
https://docs.oracle.com/javase/7/docs/technotes/guides/security/StandardNames.html#SSLContext

Note that anything that does not start with TLS is rejected in
order to prevent the weaker SSL* based protocols to be used. The
other option would be to select DTLS* but since we run on TCP
only it doesn't make sense to  use them in the first place.

Result
------
It is now possible to override the TLS version selected.

Change-Id: Ia5bbddaf03da202cfa639397796e090a96d4352c
Reviewed-on: http://review.couchbase.org/100177
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>