JVMCBC-579: RTO threshold logging in wrong order.
Motivation
----------
It was observed that, while correctly retaining the top 10 slowest
operations, the order in which they are printed is in ascending
instead of descending order (and the sdk-rfc explicitly mandates
them being printed in descending order for consistency).

Modifications
-------------
The threshold log reporter is changed (the orphan reporter after
checking doesn't need changes) so that it properly stores and
sorts the spans in descending order all the time.

Note that while adding the regression test another small issue
was found: since a set instead of a list/queue was used and the check
being done on the latency to sort the set would not insert
a span with the same duration twice. This is not a serious problme
in practice because having exactly the same latency down to the
microsecond level is very unlikely. But it can still happen and
especially during unit tests where the duration is faked it leads
to wrong assertions.

Result
------
The results are now properly sorted in descending order and a small
issue with duplicate items on the same duration has been fixed as
well.

Change-Id: I3b616c762e72b7843249efa6984823e72c760bd6
Reviewed-on: http://review.couchbase.org/100354
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>