Pick up DNS changes
Currently AbstractEndpoint caches DNS resolutions in Netty's Bootstrap
which means it can't pick up IP address changes. It will always use the
same IP address even if DNS changed. That makes it not resilent to
changes in Couchbase servers exposed via DNS names instead of IP
addresses.

A solution is to reinitialize bootstrap after a disconnect event
which should trigger a DNS request for the new IP address of the
couchbase server.