JCBC-1276 Add option to redo dns lookups for every connect
Currently AbstractEndpoint caches DNS resolutions in Netty's Bootstrap
which means it can't pick up IP address changes. It will always use the
same IP address even if DNS changed. That makes it not resilent to
changes in Couchbase servers exposed via DNS names instead of IP
addresses.

A solution is to reinitialize bootstrap after a disconnect event
which should trigger a DNS request for the new IP address of the
couchbase server.

Change-Id: I40e2b92c4ef4ce1a4faad6c485a07af7377220c9
Reviewed-on: http://review.couchbase.org/104298
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>