JVMCBC-582: Make it possible to compile on Java 11
Motivation
----------
While the set suite passes fine on Java 11, building it has failed because
there are problems with javadoc and the pegdown doclet plugin which doesn't
work anymore.

Modifications
-------------
To allow core-io being built on JDK 11, the pegdown doclet had to be removed
because it is broken and there is no update to it anymore. While it prevents
us from nicely rendering those diagrams, at least we can preserve them as
ascii for now.

Note that I had to disable the javadoc linting because javadoc throws a
tantrum with our javadoc style. I've made sure to fix the warnings with
greater than and less than at least so there are no warnings reported when
building.

Result
------
We can now build on java 11, yay!

Change-Id: I2ea21daa06b9640a73001862f3eb6112a307294c
Reviewed-on: http://review.couchbase.org/101580
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>