JVMCBC-618: Config must accept hostname like "couchbase.1".
Motivation
----------
It was reported that a hostname of "couchbase.1" lead to a view port
being identified as -1 instead of 8092. Since this is a valid hostname,
this needs to be fixed.

Modifications
-------------
URL instead of URI has no problem with a string like this that is incoming
from ns_server. I couldn't figure out why URI is swallowing a potential error
and silently returning -1 for the port instead of either accepting it (what
it should do I guess) or throwing an exception.

URL works fine and so we switch over. One KeyValueLocatorTest had to be
modified since it passed in "foo" which now also turns into a exception
(expected).

Change-Id: I5ccc5615e3a81d8b1213c28c75447503789202ec
Reviewed-on: http://review.couchbase.org/103628
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>