JVMCBC-621: Add query statement tag if present.
Motivation
----------
If tracing is used to feed data into an external system, it
is very helpful if the actual n1ql statement is included
as a tag.

Modifications
-------------
This changeset adds constructors to the request which allows
adding the statement and if it is present adds it as a tag.

Change-Id: I4de6ed28f043fd93c9bdecb1ae8101c1a78fba1b
Reviewed-on: http://review.couchbase.org/104187
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>