JVMCBC-657: Switch refresher selection logic to use the most effective one.
Motivation
----------
Right now the logic on how to select the refresher is very simplistic: if we
loaded the original config over http, it will use http for refresh. If it
loads through carrierit will use carrier for refresh.

Unfortunately, this is too simplistic: we might bootstrap off of a node in a
MDS environment where there is no KV service enabled, but throughout the
cluster there are. If it is a couchbase bucket and a modern cluster, it is more
or less a given. Using http refresh has a couple limitations and should not
be used if we can help it.

Modifications
-------------
As a result, this changeset contains a modification of the logic to select
which refresher is going to be used.

Instead of picking the same refresher as the loader, the code checks if it
is a couchbase bucket and a "modern server" (by checking for nodesExt in
the capabilities). As a result memcached buckets will still use the http
refresher, but in all other modern servers the code will now use the
carrier refresher even if we had to load through http.

Result
------
We can now switch to the carrier refresher (the most efficient and reliable
way to grab new configs) even if the user bootstrapped originally over
http, for example through a query-only node.

Change-Id: I2a65201bd0264ed70d644a0044c60c29fd425b8c
Reviewed-on: http://review.couchbase.org/108211
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>