JVMCBC-656: Only go to verbose http config if terse was not found.
Motivation
----------
The only reason to go to a verbose http config is that if the terse
one is not found: to be specific, a 404 is returned from the server.

For all other errors there is no reason to arbitrarily fall back
to verbose if terse didn't work.

This is important, because the verbose one does NOT include nodesExt
and as a result will fail to recognize services like query.

Modifications
-------------
This changeset makes sure that both in the loader and in the refresher
we will only try the verbose endpoint if the terse one failed with a
404 (which is turned into a ResponseStatus.NOT_EXISTS).

In all other cases the error is propagated which will turn into an error
in the loader and a retry attempt in the refresher case. Most importantly,
it will make sure that we will never fetch a verbose config if for some
reason the terse one failed with an error other than a not found.

Result
------
Increase the determinism and reliability when the client will fall back
to the verbose config instead of the terse one. This is important, because
the verbose one does NOT include nodesExt and as a result will fail to
recognize services like query.

Change-Id: Ic5a293cf16590cf477081c655dca44ef5ac8d179
Reviewed-on: http://review.couchbase.org/108217
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>