Restore ability to run only unit tests
Some of the unit tests referenced the TestProperties.java which
lives under the "integration test" source root. Move it under
the unit test source root so the unit tests can be run with

    mvn test -Punit
    
Doing this so Build Bot can run the unit tests again instead
of marking all our patch sets with -1.

Change-Id: Ia7cb840a903523afdb9bea09816575c759b46fca
Reviewed-on: http://review.couchbase.org/108474
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>