JVMCBC-654: Do not eagerly resolve NetworkAddresses in the config.
Motivation
----------
In order to avoid running into networking issues when dealing with
unreachable nodes (when the external address feature is used, for
example when running against a kubernetes deployment) we need
to make modifications to defer resolving to a later point in time.

Modifications
-------------
This (unfortunately rather large) changeset deals with the complex
problem in a simple way: it defers resolving by not using
NetworkAddress, and rather use the config values "as is" in their
String representation. Only when actually connecting they need
to be resolved.

Unfortunately, this change is so big because all the types change from
NetworkAddress to String, and this propagates throughout the full
stack.

Other than this pure refactoring (incl. tests)l only a couple functional
changes had to be made:

 - the service selection in core needs to take a case into account where
   there is an external address, but the ports are still used from the
   "regular" section.
 - the optional eager DNS resolve logic has been removed, since it is not
   needed anymore and would try to resolve a potential hostname that
   it can't reach, defeating the purpose of this change.
 - IPv6 parsing has to be expanded to remove prefix and suffix [] which
   got handled by the underlying NetworkAddress in the past.

Note that this only includes the minimum type changes that had to be made
to get this wokring. I'm sure there are more places where can rip out the
NetworkAddress usage, but this can be done in follow up changesets.

Change-Id: Ia071182af38ca2047aa352b0304edcd1dd24a044
Reviewed-on: http://review.couchbase.org/109685
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>