JVMCBC-677: Fix division by 0 in ConfigLocator
Motivation
----------
During shutdown there is a race condition which might trigger
a division by 0 in the config locator. It is not particularly
harmful, but still it needs to be cleaned up.

Modifications
-------------
the ConfigLocator now properly checks if the node list is empty,
and if it is performs a retry of the op. All the other locators
work as expected, but just in case regression unit tests have been
added.

Change-Id: Ide85266d270c115046ae10cdadb6d7601a8d1079
Reviewed-on: http://review.couchbase.org/110030
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>