JVMCBC-654: NetworkAddress cleanup in core.
Motivation
----------
This changeset removes the imports and remaining usage of NetworkAddress
throughout the codebase.

Note that it adds a usage in the endpoint which is explained below.

Modifications
-------------
The unused import cleanup is not very exciting, although the important
piece is that this changeset adds the use of the NetworkAddress inside
the AbstractEndpoint. The reasoning behind this is that we need to
preserve the edge case where if we are running in IPv6 jvm and using
an old cluster we need to force a IPv4 lookup. This codepath is only
used if the flag is set though, which the user needs to enable manually.

Also, in this spot it is safe to use since we connect right afterwards
so the hostname needs to be reachable somehow.

Change-Id: Ic9f53615dc93e1d69a9d84077242d93cbbff0886
Reviewed-on: http://review.couchbase.org/110034
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>