JVMCBC-702 Test failures when using mock
Motivation
==========
We run the tests against the mock, and all of them should pass.  A couple
don't, due to limitations in the mock.  Also, -DuseMock=true|false should
work, and seems not to.  That's bad - complicates the interpretation of any
integration tests, etc...

Modification
===========
The 2 tests need to be aware that a mock is being used, and not ask for things
the mock doesn't support.  Also, we expose useMock as a property in the pom, which
the maven resources plugin then pops in the expected spot in the properties file.  This
allows -DuseMock to work.

Change-Id: Ic6e1399ba3aa699b4642b5932c922fc471d7689c
Reviewed-on: http://review.couchbase.org/111756
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>