JVMCBC-712: Update seedHosts with every good new config.
Motivation
----------
Previously core-io would not update the seed nodes once a good config
has been found, but rather the original one would be used. This worked fine
since openBucket attempts are usually made when the app starts and only
over time the config list would diverge if nodes are added/removed from
the cluster.

A recent modification to support Kubernetes deployments which in turn made
it impossible to resolve hostnames early lead to a regression which manifests
as follows:

Modifications
-------------

 - Every time an open bucket attempt is made, the seed hosts are used to
   dispatch located configs to that specific nodes.
 - If the client uses ip addrs and the server hostnames, or vice versa,
   after the second bucket open attempt this hostname might not be dispatchable
   anymore since the new config already arrived, leaving subsequent openBucket
   attempts unresolvable.
 - Previously this wasn't an issue since the ip and host resolved to the same
   address, but now we need to make sure that the seedHosts is always up-to-date
   with the latest configs so always the right seed names are dispatched.

This also has the added benefit that should someone open a bucket later in the process
and the topology changes, we'll still be able to bootstrap properly.

Change-Id: Ibcfa2c68629fa7a8d1ca9dca4f14a9cb60b8f723
Reviewed-on: http://review.couchbase.org/112560
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>