JVMCBC-713: Do not try to resolve in ConnectionString
Motivation
----------
Not only do we need to avoid resolving early in the core, also for
the actual bootstrap list nodes we need to make sure that they are
not being resolved. With this change the hostname or ips are passed
through 1:1 as they are provided, making sure that the seed host
list is deterministic and can be debugged properly in the logs.

NOTE: this is technically a breaking change for anyone relying on the
hosts() or allHosts() return type. Since this API is used internally
though and not being exported, I do not expect breakage of any sorts,
and if so the change to fix is minimal (see test cases for examples).

Change-Id: I6abf1510ea45dbbc1f0fdd4e0173d160a859725f
Reviewed-on: http://review.couchbase.org/112563
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>