JVMCBC-714: Do not resolve prelocated requests.
Motivation
----------
Since all the hostnames need to match equally now, we cannot resolve
the prelocated requests (i.e. always into IPs) in the query locator.
This might cause undispatchable requests if the hostnames are actual
hostnames and not IPs.

Modifications
-------------
The prelocated request has been modified so it returns a String and
all the implementations that use it have been modified so they pass
in the string and not try to stick it into a InetAddr.

Result
------
Prelocated requests can now be dispatched again.

Change-Id: I8188ecd8c3c46240a3bf69295d04eda701b526b4
Reviewed-on: http://review.couchbase.org/112564
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>