JVMCBC-715: Fix analytics keepalive for new handler.
Motivation
----------
When creating the V2 handler, the keepalive request for analytics
(since it is created as part of the instance) lives inside the handler.
When copying the implementation over for v2, it still referenced the
v1 handler, which the encode method doesn't accept. This leads to keepalives
failing for analytics when this handler is used (which it is by default).

Change-Id: I2334561a515d09ec32793b8d907707d2e7c68d2b
Reviewed-on: http://review.couchbase.org/112566
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>