SDKQE-289 Deploy Snapshots
Motivation
==========
Integration tests, when successful, should publish snapshots to
maven central.

Modification
===========
This really already happens.  However, they are not signed.  The
pom needed a bit of modification to sign - once this is in Jenkins
needs the appropriate settings and we will be signing.  Also there
is a bug in one of our tests, which I fixed here.

Change-Id: Ifc17856d30a5068f72231ac733756d0c8f02ba0e
Reviewed-on: http://review.couchbase.org/112813
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>