JVMCBC-748: Seed hosts should only include KV nodes on update
Motivation
----------
2.7.9 introduced a feature which keeps the seed nodes up-to-date
as the configuration state changes. Previously it added all the
hostnames, but it should only add the nodes which have the KV
service enabled in the first place.

This avoids the situation where subsequent bucket open attempts
will reach nodes that do not have KV enabled and as a result have
"ugly" warnings in the logs - although not considered harmful at
this point.

Change-Id: Iee4b353b2b9dccb788e7f0a4e37f03a088679637
Reviewed-on: http://review.couchbase.org/115553
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>