Add OpenSSL support to couchbase client. Motivation ---------- Couchbase Core IO uses netty4 as base. Netty4 supports OpenSSL. OpenSSL provides better performance than the native JDK implementation. This commit tries to enable openSSL by using the boring ssl implementation of tc native available for Netty4.
By default openSSL is disabled in the couchbase client SSL factory. The feature can be enabled by explicitly
overriding the property openSslEnabled in the DefaultEnvironment builder.

Limitation: The current couch base SSLEngineFactory class allows ssl protocols to be explicitly configured using
the com.couchbase.sslProtocol property. This was introduced to configure TLS properties and also limit the use
of the protocols to TLS while disabling weaker protocols like SSL. Given that netty version is 4.0.56.Final, we
cannot enforce the protocols to only TLS. This bug is documented https://github.com/netty/netty/issues/7935 and
has been fixed in subsequent versions. So when the openSSL implementation is enabled, it is not possible to disable
SSL protocols explicitly.