JVMCBC-759: Support TLS Hostname Verification
Motivation
----------
This changeset adds support for hostname verification if explicitly
enabled on the environment. Note that it is not enabled by default
since users of previous versions likely rely in their setup on
the fact that hostname verification was not available.

Modifications
-------------
When enabled on the SecureEnvironment, the SSLEngineFactory applies
the endpointIdentificationAlgorithm of HTTPS (available sinde Java 7).

Result
------
If explicitly enabled, provides hostname verification support for
tls.

Change-Id: If23c2cc2e4e557fbcfcf333eb7c1f1d9ae86a8e9
Reviewed-on: http://review.couchbase.org/117670
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>