JVMCBC-807: Hostnames are not re-resolved after initial bootstrap
In the netty bootstrap logic, when the remote hostname is added
it resolves to an inet address. During reconnect attempts if the
bootstrap is cached the host is never re-resolved so if the DNS
changes this might lead to issues.

The fix is that we just re-apply the same hostname on every
connect attempt which will make sure that it is resolved
fresh each time.

Change-Id: I087f5fafbfeb805b0cd463a86eeb519c5366f572
Reviewed-on: http://review.couchbase.org/120650
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>