JVMCBC-844: Null out request and response in json parser after completion.
This changeset makes sure that when a request is completed in the JSON parser
both the referenced request and response are nulled out since we do not
need to access it anymore.

This has the positive side effect that the GC has an easier job cleaning up
outstanding resources for the last response even if no new request is
incoming (since if a new one comes in the variables are reassigned).

Change-Id: I5b003eb4b21847584303b1e0493e8fa2a3eb1836
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/125771
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>