JVMCBC-854: yasjl parser incorrectly publishes a truncated number with > 2 bytes as a result
In readValue, a call to content.forEachByte() returns -1 if the end of
the readable stream was hit while parsing the bytes.

In almost all cases this causes the parser to throw an exception
requesting more data.

However, for some reason, if it's a number, and if more than 2 readable
bytes were present in the buffer, it decides to instead process it as a
valid number.  E.g. it raises the truncated value as a row.

This also appears to finish the row processing.

The change is to remove the special-case handling for these numbers, whose
purpose is unclear.

Manual testing confirms that the bug is resolved, with a stream of numbers
now processed correctly, regardless of how that stream is split up in
packets.

Change-Id: Idb9d6915dbd56cce46f76c3fa65a5feedade5007
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/129518
Tested-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>