JVMCBC-900 - Reduce default http idle time to 4s
This change brings the default http idle time down to 4s
and also increases the check interval considerably so we
make sure that idle http sockets are definitely closed before
the new server 5s http idle timeout hits us.

Note that 3.x has it set to 4.5s and here it is 4s. The
reason for this is that the unit used is seconds and I
did not want to change/break the api, and it should be
close enough for those purposes.

Change-Id: If10a8034b2e0570e689e9f97b97fc404bb921bee
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/139362
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>