JVMCBC-902: Properly handle sockets which have no request sent yet in idle cleaner
Motivation
----------
1.7.17 surfaced an issue which has been in the code base for a long time, but
has never really shown up. Now with the decreased idle time poller at 100ms,
this race condition shows up and is an issue for pooled operations, like
query requests.

Modifications
-------------
The issue is that the poller only took the last response received timestamp
into account when deciding if it can be cleaned up or not. But if the socket
is still being bootstrapped and/or has not had a chance to get a full
response back yet, the socket would be closed very quickly and might never
even had a chance to perform a single op. This issue comes up even more in
slower network scenarios, i.e. when connecting to the cloud from a local
network.

The fix for this issue is the same that we already have in place for
SDK 3: In addition to the response received timestamp, we introduce
another one: when the socket has been connected in the first place.

If that time is 0, the socket is not even connected and we must not
clean it up. If it is != 0 but no response is received yet, calculate
the idle time based off of it. Once we have a response, the code will
behave like the old code did before.

Note: it is not possible that we'll "shoot" an ongoing request, because
the code also checks if the endpoint is free (which it is not when
a request is outstanding).

Result
------
Even with the quick check intervals, endpoints are not prematurely
disconnected and never get a chance to execute the request in the
first place.

Change-Id: I3abda21daa7f43fde80f8969ec72dd9c603f5624
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/140116
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>