JVMCBC-903: Also handle HTTP 409
It has been reported that the query engine can return a
HTTP 409 when an index is already present, and SDK 2 does
not handle this status code and emits a warning which is
noisy.

This changeset maps the 409 code to the closest equivalent
(not stored) since it shows up during PUT operations indicating
that it could not store it due to a conflict.

Change-Id: I4010318aa62ec7d58a3bcb7aabccb4a1433efba9
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/141140
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>