JVMCBC-923: Use hostname from config for HTTP host header
This changeset aligns with a recent SDK 3 change where instead
of always using the IP address for the http host header this
change now makes sure that we use whatever hostname/ip is part
of the server config that we recieve. This makes sure that we
also pass the same http host header that we used for originally
connecting to the socket.

Note that the old code is still there as a fallback since in mocked
tests we do not have access to the hostname and this change has
been designed to have minimal impact on the surrounding code.

Change-Id: Ie458ef9f0e69e3c272addaee04c29df5a9fb4e55
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/146054
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>