JVMCBC-945: Explicitly handle UnknownHostException
This changeset explicitly handles a UnknownHostException when raised
from inside netty. The connect fails anyways, but this makes sure that
the calling Observable is notified properly in this case.

As a drive by, a message has been upped to a warning when raised and
some unnecessary throws Exception blocks have been removed.

Change-Id: I260f0e0cffc3360bb9a53325971efe4e0ee0b1e7
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/148688
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>