JVMCBC-953: Use FF map only if NMVB at least present once
Motivation
----------
Before this change, during a rebalance (if a FF map is present
in the server config), the client would try the fast forward
node even if the operation got retried for a different reason
other than a NMVB in the first place.

While it is true that moste of the time in this case it could
be a NMVB, it doesn't have to be all the time. Since the server
expects us to ONLY try an operation on the FF node if a NMVB
got returned, the logic needs to be adjusted.

Modifications
-------------
Instead of checking on the number of retries (> 0), this code
change now sets a flag on the request if a NMVB response has
been received and only then it will try to attempt to use the
fast forward map.

Change-Id: Id4b69ee482532048bddf20bc96c639ded7e60143
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/150078
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>